requirements.txt
# Core Django
Django==5.2.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
gunicorn==23.0.0

# Database/Storage
boto3==1.38.41
django-storages==1.14.6
redis==6.2.0
django-redis==6.0.0

# Security
cryptography==44.0.2
pyOpenSSL==25.0.0

# Utilities
python-decouple==3.8
python-dotenv==1.1.0
whitenoise==6.6.0

# Optional (only include if used)
tensorflow_cpu==2.19.0  # Only if AI features need it
geopy==2.4.1  # Only if location services needed
razorpay==1.4.2  # Only if payment processing needed
